gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
redis[asyncio]==5.0.4
httpx==0.27.0
python-multipart==0.0.9
structlog==24.1.0
python-dotenv==1.0.1
